Arm backend: Move debug related functions to common/debug.py
Move debug functions from tosa_utils.py to common/debug.py. It makes
more sense since the debug functions aren't just for tosa information as
well as making the import simpler with an upcoming move of tosa files to
a separate tosa directory in the Arm backend.
